% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xfaostat_helper_funcs.R
\name{assert_FBS_balance}
\alias{assert_FBS_balance}
\title{assert processed food balance sheet or supply utilization account is balanced}
\usage{
assert_FBS_balance(.DF)
}
\arguments{
\item{.DF}{Input data frame}
}
\value{
massages or warnings
}
\description{
assert processed food balance sheet or supply utilization account is balanced
}
