% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xfaostat_visualization_funcs.R
\name{chord_wrapper}
\alias{chord_wrapper}
\title{chord_wrapper: a wrapper function of the chordDiagram function
Title}
\usage{
chord_wrapper(
  .DF,
  .FigTitle = NULL,
  .SaveDir = "../man/figures",
  .GRIDCOLOR = NULL,
  .ORDER = NULL,
  .SaveName,
  .SaveScaler = 1
)
}
\arguments{
\item{.DF}{Input data frame.}

\item{.FigTitle}{Figure title}

\item{.SaveDir}{The directory for saving the plot; the default is man/figures.}

\item{.GRIDCOLOR}{Colors of grid; consider using circlize::rand_color function}

\item{.ORDER}{Order of the gird}

\item{.SaveName}{Name of the saved file in .SaveDir}

\item{.SaveScaler}{A scaler controling the size (dpi = 600)}
}
\value{
A saved plot in .SaveDir
}
\description{
chord_wrapper: a wrapper function of the chordDiagram function
Title
}
