% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xfaostat_helper_funcs.R
\name{FAO_AREA_RM_NONEXIST}
\alias{FAO_AREA_RM_NONEXIST}
\title{FAO_AREA_RM_NONEXIST: remove nonexistent FAO region using area_code, e.g., USSR after 1991}
\usage{
FAO_AREA_RM_NONEXIST(
  .DF,
  SUDAN2012_MERGE = F,
  RM_AREA_CODE = c(69, 87, 127, 135, 145, 182, 299)
)
}
\arguments{
\item{.DF}{A input data frame}

\item{SUDAN2012_MERGE}{If TRUE, only keep one merged Sudan region}

\item{RM_AREA_CODE}{Additional area_codes to remove, e.g., small regions with low data quality.}
}
\value{
A data frame with nonexistent area and area_code removed
}
\description{
This function was developed when exploring FAO production data.
Additional small regions/areas are also remove due to low data quality.
}
