% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xfaostat_helper_funcs.R
\name{FAOSTAT_metadata}
\alias{FAOSTAT_metadata}
\title{FAOSTAT_metadata: get the most recent metadata from FAOSTAT API and return a summary data frame}
\usage{
FAOSTAT_metadata(code = NULL)
}
\arguments{
\item{code}{FAOSTAT dataset code if filtering specific dataset; NULL by default and return all dataset info}
}
\value{
Dataframe including the metadata of FAOSTAT including URL for download, updating date, detailed descriptions, etc.
}
\description{
Download and parse the metadata file from the FAOSTAT portal. FAOSTAT_metadata was adapted from the FAOSTAT package.
}
\examples{
FAOSTAT_metadata
}
