% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xfaostat_helper_funcs.R
\name{FF_check_count_plot}
\alias{FF_check_count_plot}
\title{FF_check_count_plot: count item_code and area_code by year}
\usage{
FF_check_count_plot(.DF, .ELEMENT = c())
}
\arguments{
\item{.DF}{Input data frame}

\item{.ELEMENT}{A set of elements (in Char) to focus. If empty, all elements are summarized}
}
\value{
A plot summarizing the time-series of changing the count of item_code and area_code (grouped by element).
}
\description{
FF_check_count_plot: count item_code and area_code by year
}
