% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xfaostat_helper_funcs.R
\name{FF_download_FAOSTAT}
\alias{FF_download_FAOSTAT}
\title{FF_download_FAOSTAT: Bulk download raw data from FAOSTAT API by dataset code}
\usage{
FF_download_FAOSTAT(
  DATASETCODE,
  DATA_FOLDER = file.path("inst/extdata", DIR_RAW_DATA_FAOSTAT),
  OverWrite = FALSE
)
}
\arguments{
\item{DATASETCODE}{Single dataset code in FAO metadata, e.g., QCL is the production dataset.}

\item{DATA_FOLDER}{Destination folder for storing raw data}

\item{OverWrite}{If FALSE, check if the dataset exists and stop if exists; overwrite if TRUE}
}
\description{
FF_download_FAOSTAT: Bulk download raw data from FAOSTAT API by dataset code
}
