% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xfaostat_helper_funcs.R
\name{FF_download_faostat_bulk}
\alias{FF_download_faostat_bulk}
\title{FF_download_faostat_bulk: Bulk download raw data from FAOSTAT API by dataset code}
\usage{
FF_download_faostat_bulk(DATASETCODE, DATA_FOLDER = DIR_RAW_DATA_FAOSTAT)
}
\arguments{
\item{DATASETCODE}{Dataset code in FAO metadata, e.g., QCL is the production dataset.}

\item{DATA_FOLDER}{Destination folder for storing raw data}
}
\description{
FF_download_faostat_bulk: Bulk download raw data from FAOSTAT API by dataset code
}
