% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xfaostat_helper_funcs.R
\name{FF_download_RemoteArchive}
\alias{FF_download_RemoteArchive}
\title{FF_download_RemoteArchive: Download raw data from remote archive (Zenodo)}
\usage{
FF_download_RemoteArchive(
  DATASETCODE = NULL,
  RemoteArchiveURL = "https://zenodo.org/record/8260225/files/",
  DATA_FOLDER = DIR_RAW_DATA_FAOSTAT,
  OverWrite = FALSE
)
}
\arguments{
\item{DATASETCODE}{Single dataset code in FAO metadata, e.g., QCL is the production dataset.}

\item{RemoteArchiveURL}{Zenodo URL, default (GCAM v7) = "https://zenodo.org/record/8260225/files/"}

\item{DATA_FOLDER}{Folder stores raw data; default sets to DIR_RAW_DATA_FAOSTAT}

\item{OverWrite}{If FALSE, check if the dataset exists and stop if exists; overwrite if TRUE}
}
\description{
FF_download_RemoteArchive: Download raw data from remote archive (Zenodo)
}
