% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xfaostat_helper_funcs.R
\name{FF_FILL_NUMERATOR_DENOMINATOR}
\alias{FF_FILL_NUMERATOR_DENOMINATOR}
\title{FF_FILL_NUMERATOR_DENOMINATOR Fill in missing values considering relationship between two variables}
\usage{
FF_FILL_NUMERATOR_DENOMINATOR(
  .DF,
  NUMERATOR_c,
  DENOMINATOR_c,
  NUMERATOR_FILL_DIRECTION = "down"
)
}
\arguments{
\item{.DF}{Input data frame}

\item{NUMERATOR_c}{NUMERATOR col name (production) in the ratio}

\item{DENOMINATOR_c}{DENOMINATOR col name (area) in the ratio}

\item{NUMERATOR_FILL_DIRECTION}{Direction of filling in NUMERATOR}
}
\value{
A data frame with missing values filled
}
\description{
Fill in based on relationship between NUMERATOR_c and DENOMINATOR_c.
E.g., yield = production / area. NUMERATOR_c will be linearly interpolated forward
then NUMERATOR_FILL_DIRECTION. Yield will be filled down-up, World average yield
is used when only NUMERATOR_c or DENOMINATOR_c is available
}
