% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xfaostat_helper_funcs.R
\name{FF_rawdata_info}
\alias{FF_rawdata_info}
\title{FF_rawdata_info: extract meta info of data (e.g., zip files) in a folder}
\usage{
FF_rawdata_info(DATA_FOLDER = DIR_RAW_DATA, DATASETCODE, DOWNLOAD_NONEXIST = F)
}
\arguments{
\item{DATA_FOLDER}{The folder including raw zipped csv data from FAOSTAT.}

\item{DATASETCODE}{Dataset code in FAO metadata, e.g., QCL is the production dataset.}

\item{DOWNLOAD_NONEXIST}{A logical variable. If TRUE, nonexist dataset will be downloaded.}
}
\value{
A data frame summarizing the local raw data info (e.g., size, downloaded date, etc.) if DOWNLOAD_NONEXIST == FALSE.
Otherwise, nonexistent files will be downloaded and the function should be rerun.
}
\description{
FF_rawdata_info: extract meta info of data (e.g., zip files) in a folder
}
