% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xfaostat_helper_funcs.R
\name{FF_summary}
\alias{FF_summary}
\title{FF_summary summarize a loaded FAO dataset}
\usage{
FF_summary(DF, COL_CNTY = "area", COL_ITEM = "item", .ENVIR = .GlobalEnv)
}
\arguments{
\item{DF}{A data frame}

\item{COL_CNTY}{Country column name}

\item{COL_ITEM}{Item column name}

\item{.ENVIR}{Environment of the DF}
}
\value{
A data frame with summary information,
}
\description{
FF_summary summarize a loaded FAO dataset
}
