% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xfaostat_helper_funcs.R
\name{gcamfaostat_metadata}
\alias{gcamfaostat_metadata}
\title{gcamfaostat_metadata: generate metadata information of the input data}
\usage{
gcamfaostat_metadata(
  .DIR_RAW_DATA_FAOSTAT = DIR_RAW_DATA_FAOSTAT,
  OnlyReturnDatasetCodeRequired = FALSE
)
}
\arguments{
\item{.DIR_RAW_DATA_FAOSTAT}{Path to raw faostat data; default set to DIR_RAW_DATA_FAOSTAT in constants.R}

\item{OnlyReturnDatasetCodeRequired}{If true only required dataset codes are returned}
}
\value{
Information of FAOSTAT input dataset
}
\description{
gcamfaostat_metadata: generate metadata information of the input data
}
