% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zchunk_LA100.FAO_SUA_PrimaryEquivalent.R
\name{module_aglu_LA100.FAO_SUA_PrimaryEquivalent}
\alias{module_aglu_LA100.FAO_SUA_PrimaryEquivalent}
\title{module_aglu_LA100.FAO_SUA_PrimaryEquivalent}
\usage{
module_aglu_LA100.FAO_SUA_PrimaryEquivalent(command, ...)
}
\arguments{
\item{command}{API command to execute}

\item{...}{other optional parameters, depending on command}

}
\value{
Depends on \code{command}: either a vector of required inputs, a vector of output names, or (if
  \code{command} is "MAKE") all the generated outputs: \code{GCAM_AgLU_SUA_APE_2010_2019},
  \code{FAO_AgProd_Kt_All},\code{FAO_AgArea_Kha_All},\code{FAO_Food_Macronutrient_All_2010_2019},
  \code{FAO_Food_MacronutrientRate_2010_2019_MaxValue}

Trade-adjusted supply-utilization accounting data frame
Regional SUA aggregation (e.g., 32 in GCAM) with an option of removing intra-regional trade

Check and assert SUA data frame structure

Logical T or F
Get extraction rate

A data frame including regional, traded, and world extraction rates of a processing
Separate the SUA balance into domestic and imported balanced for sink_item

SUA DF
Separate the domestic SUA balance into current and lagged balanced for sink_item

SUA DF
Primary equivalent aggregation

A supply-utilization accounting data frame of a data tier in source item equivalent
Recursively aggregate nest into primary equivalent using from bottom to top

SUA data frame in primary equivalent
}
\description{
The function is used as an option in SUA_REG_Agg

intra-regional trade, by default, removed only for the SUA items with bilateral trade data available.
E.g., for 32 regions, if removed, global trade becomes inter-32-region trade.

Gross extraction rate is calculated for domestic, traded, and lagged values.
By gross, it means sink items are aggregated.
The function is used in Proc_primarize.

The function is used in Proc_primarize

The function is used in Proc_primarize

The function is a recursive wrapper of Proc_primarize
}
\details{
Generate supply utilization balance in primary equivalent

This chunk compiles balanced supply utilization data in primary equivalent in GCAM region and commodities.
A method to generate primary equivalent is created for the new FAOSTAT supply utilization data (2010 to 2019).
New SUA balance is connected to the old one (before 2010). Production and harvested area data with FAO region and item
for primary production are provided. For FAO food items, macronutrient values are calculated at SUA item level.
Data processing was consistent across scales. Note that GCAM regions and commodities in aggregation mapping can
be changed in corresponding mappings. The output data is not averaged over time.
}
\author{
XZ 2022
Adjust gross trade in SUA data to ensure regional export is smaller than production for an SUA item
}
