% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xfaostat_L102_QCLProductionArea.R
\name{module_xfaostat_L102_QCLProductionArea}
\alias{module_xfaostat_L102_QCLProductionArea}
\title{module_xfaostat_L102_QCLProductionArea}
\usage{
module_xfaostat_L102_QCLProductionArea(command, ...)
}
\arguments{
\item{command}{API command to execute}

\item{...}{other optional parameters, depending on command}
}
\value{
Depends on \code{command}: either a vector of required inputs, a vector of output names, or (if
  \code{command} is "MAKE") all the generated outputs
}
\description{
Preprocess raw faostat data
}
\details{
This chunk compiles balanced supply utilization data in primary equivalent in GCAM region and commodities.
}
\author{
XZ 2023
}
