% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xfaostat_L999_DataExport.R
\name{module_xfaostat_L999_DataExport}
\alias{module_xfaostat_L999_DataExport}
\title{module_xfaostat_L999_DataExport}
\usage{
module_xfaostat_L999_DataExport(command, ...)
}
\arguments{
\item{command}{API command to execute}

\item{...}{other optional parameters, depending on command}
}
\value{
Depends on \code{command}: either a vector of required inputs, a vector of output names, or (if
  \code{command} is "MAKE") all the generated outputs
}
\description{
Generate supply utilization balance in primary equivalent
}
\details{
This chunk compiles balanced supply utilization data in primary equivalent in GCAM region and commodities.
}
\author{
XZ Sep2024
}
