% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yextension_L100_FoodBalanceSheet.R
\name{module_yextension_L100_FoodBalanceSheet}
\alias{module_yextension_L100_FoodBalanceSheet}
\title{module_yextension_L100_FoodBalanceSheet}
\usage{
module_yextension_L100_FoodBalanceSheet(command, ...)
}
\arguments{
\item{command}{API command to execute}

\item{...}{other optional parameters, depending on command}
}
\value{
Depends on \code{command}: either a vector of required inputs, a vector of output names, or (if
  \code{command} is "MAKE") all the generated outputs: \code{FAO_Food_Macronutrient_All_2010_2019},
  \code{FAO_Food_MacronutrientRate_2010_2019_MaxValue}
}
\description{
Generate supply utilization balance in primary equivalent
}
\details{
This chunk aggregates supply utilization accounts into food balance sheet items
}
\author{
XZ 2024
# In some cases, prescale sink item SUA using output_specific_extraction_rate can improve the processing.
# e.g., when coproduction shares are not fixed.
}
