% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xfaostat_helper_funcs.R
\name{output_csv_data}
\alias{output_csv_data}
\title{Function saving dataset to csv file with headers}
\usage{
output_csv_data(
  gcam_dataset,
  col_type_nonyear,
  title,
  unit,
  description = NA,
  code,
  out_dir = out_dir,
  GZIP = F
)
}
\arguments{
\item{gcam_dataset}{dataframe name to be saved}

\item{col_type_nonyear}{column type that is non-year; numeric for years will be pasted}

\item{title}{title in header}

\item{unit}{unit in header}

\item{description}{description in header}

\item{code}{fao dataset domain code}

\item{out_dir}{output directory}

\item{GZIP}{IF TRUE}
}
\description{
Function saving dataset to csv file with headers
}
