% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{PREBUILT_DATA}
\alias{PREBUILT_DATA}
\title{A list of prebuilt data objects. FAOSTAT raw data, those generated in modules
xfaostat_L101_RawDataPreProc*, are included to facilitate the processing.
Its immediate downstream dependencies then used the prebuilt versions of
their outputs stored in this object.}
\format{
A list object where [[object_name]] <- tibble:
}
\usage{
PREBUILT_DATA
}
\description{
A list of prebuilt data objects. FAOSTAT raw data, those generated in modules
xfaostat_L101_RawDataPreProc*, are included to facilitate the processing.
Its immediate downstream dependencies then used the prebuilt versions of
their outputs stored in this object.
}
\keyword{datasets}
