% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xfaostat_helper_funcs.R
\name{rm_accent}
\alias{rm_accent}
\title{rm_accent: Remove accent and prime in selected columns of a data frame}
\usage{
rm_accent(.df, ...)
}
\arguments{
\item{.df}{Input data frame}

\item{...}{A character set of column names}
}
\value{
A data frame with accent and prime removed
}
\description{
rm_accent: Remove accent and prime in selected columns of a data frame
}
