% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xfaostat_helper_funcs.R
\name{SUA_bal_adjust}
\alias{SUA_bal_adjust}
\title{Balance Supply Utilization Accounting data}
\usage{
SUA_bal_adjust(.df)
}
\arguments{
\item{.df}{Input SUA data frame}
}
\value{
SUA dataset that is balanced across different dimensions
}
\description{
Be careful with the unit as value divided by 1000 here; Light adjustments to merge Tourist consumption into food
}
