% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xfaostat_visualization_funcs.R
\name{treemap_wrapper}
\alias{treemap_wrapper}
\title{treemap_wrapper: a warpper function of the treemap function}
\usage{
treemap_wrapper(
  .DF,
  .Depth = 2,
  .Palette = "Set2",
  .FigTitle,
  .FigTitleSize = 12,
  .LastLabelCol = "orange",
  .SaveDir = "../man/figures",
  .SaveName,
  .SaveScaler = 1
)
}
\arguments{
\item{.DF}{Input data frame. The data frame needs to include index columns (need to be first columns) and a value column.}

\item{.Depth}{Grouping depth of the tree; the columns will be used. The first 2 columns in .DF will be index in treemap}

\item{.Palette}{Color palette (RColorBrewer is used)}

\item{.FigTitle}{Figure title}

\item{.FigTitleSize}{Size of title font}

\item{.LastLabelCol}{Color of the last label}

\item{.SaveDir}{The directory for saving the plot; the default is man/figures.}

\item{.SaveName}{Name of the saved png file in .SaveDir}

\item{.SaveScaler}{A scaler controling the size (dpi = 300)}
}
\value{
A saved plot in .SaveDir
}
\description{
See https://cran.r-project.org/web/packages/treemap/index.html for the details of the treemap function
}
