% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xfaostat_helper_funcs.R
\name{gcamfaostat_metadata}
\alias{gcamfaostat_metadata}
\title{gcamfaostat_metadata: generate metadata information of the input data}
\usage{
gcamfaostat_metadata(OnlyReturnDatasetCostRequired = FALSE)
}
\arguments{
\item{OnlyReturnDatasetCostRequired}{If TRUE only return dataset codes set in the function}
}
\value{
Information of FAOSTAT input dataset
}
\description{
gcamfaostat_metadata: generate metadata information of the input data
}
