% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zchunk_LA100.FAO_SUA_PrimaryEquivalent.R
\name{module_aglu_LA100.FAO_SUA_PrimaryEquivalent}
\alias{module_aglu_LA100.FAO_SUA_PrimaryEquivalent}
\title{module_aglu_LA100.FAO_SUA_PrimaryEquivalent}
\usage{
module_aglu_LA100.FAO_SUA_PrimaryEquivalent(command, ...)
}
\arguments{
\item{command}{API command to execute}

\item{...}{other optional parameters, depending on command}

\item{.RM_IntraRegTrade}{If TRUE, intra-regional trade is removed after aggregation}

\item{.TM_Bilateral}{A data frame of bilateral trade data}

\item{.RM_TRIANGLE_INEQUALITY}{If TRUE, adjust gross trade in SUA data to ensure regional export is smaller than production for an SUA item}

\item{.source_item}{Source items or primary items in the processing}

\item{.sink_item}{Sink items or processed items in the processing}

\item{.SOURCE_ITEM}{Source items or primary items in the processing}

\item{.SINK_ITEM}{Sink items or processed items in the processing}

\item{.OUTPUT_SPECIFIC_EXTRACTION_RATE}{A data frame with item and output_specific_extraction_rate.
# In some cases, prescale sink item SUA using output_specific_extraction_rate can improve the processing.
# e.g., when coproduction shares are not fixed.}

\item{.COMM}{Primary commodity in the FAO item mapping (APE_comm).}

\item{.MAPPING}{Mapping of SUA items to primary products by level of nestings. Mapping_SUA_PrimaryEquivalent was constructed.}

\item{.DF_SUA}{Input SUA data frame}

\item{.ITEM_SUFFIX}{Adding a suffix (_APE) for item in primary equivalent in output. Note that an aggregated primary commodity is also kept.}

\item{.RECUR_TOP_NEST}{Bottom nest_level in mapping. If 1, all the way through APE but if 2 only run to top but 1 nest}

\item{.DF}{An input dataframe with an element col including Import and Export}

\item{.MIN_TRADE_PROD_RATIO}{Trade will be removed if world total export or import over production is smaller than .MIN_TRADE_PROD_RATIO (1% default value)}

\item{.Reg_VAR}{Region variable name; default is ("area_code")}

\item{.GROUP_VAR}{Group variable; default is ("item_code", "year")}
}
\value{
Depends on \code{command}: either a vector of required inputs, a vector of output names, or (if
  \code{command} is "MAKE") all the generated outputs: \code{GCAM_AgLU_SUA_APE_1973_2019},
  \code{FAO_AgProd_Kt_All},\code{FAO_AgArea_Kha_All},\code{FAO_Food_Macronutrient_All_2010_2019},
  \code{FAO_Food_MacronutrientRate_2010_2019_MaxValue}

Trade-adjusted supply-utilization accounting data frame
Regional SUA aggregation (e.g., 32 in GCAM) with an option of removing intra-regional trade

Check and assert SUA data frame structure

Logical T or F
Get extraction rate

A data frame including regional, traded, and world extraction rates of a processing
Separate the SUA balance into domestic and imported balanced for sink_item

SUA DF
Separate the domestic SUA balance into current and lagged balanced for sink_item

SUA DF
Primary equivalent aggregation

A supply-utilization accounting data frame of a data tier in source item equivalent
Recursively aggregate nest into primary equivalent using from bottom to top

SUA data frame in primary equivalent
Balance gross trade

The same dataframe with balanced world export and import.
}
\description{
The function is used as an option in SUA_REG_Agg

intra-regional trade, by default, removed only for the SUA items with bilateral trade data available.
E.g., for 32 regions, if removed, global trade becomes inter-32-region trade.

Gross extraction rate is calculated for domestic, traded, and lagged values.
By gross, it means sink items are aggregated.
The function is used in Proc_primarize.

The function is used in Proc_primarize

The function is used in Proc_primarize

The function is a recursive wrapper of Proc_primarize

Scale gross export and import in all regions to make them equal at the world level.
}
\details{
Generate supply utilization balance in primary equivalent

This chunk compiles balanced supply utilization data in primary equivalent in GCAM region and commodities.
A method to generate primary equivalent is created for the new FAOSTAT supply utilization data (2010 to 2019).
New SUA balance is connected to the old one (before 2010). Production and harvested area data with FAO region and item
for primary production are provided. For FAO food items, macronutrient values are calculated at SUA item level.
Data processing was consistent across scales. Note that GCAM regions and commodities in aggregation mapping can
be changed in corresponding mappings. The output data is not averaged over time.
}
\author{
XZ 2022
Adjust gross trade in SUA data to ensure regional export is smaller than production for an SUA item
}
