% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xfaostat_L105_DataConnectionToSUA.R
\name{module_xfaostat_L105_DataConnectionToSUA}
\alias{module_xfaostat_L105_DataConnectionToSUA}
\title{module_xfaostat_L105_DataConnectionToSUA}
\usage{
module_xfaostat_L105_DataConnectionToSUA(command, ...)
}
\arguments{
\item{command}{API command to execute}

\item{...}{other optional parameters, depending on command}
}
\value{
Depends on \code{command}: either a vector of required inputs, a vector of output names, or (if
  \code{command} is "MAKE") all the generated outputs
}
\description{
Connect datasets to build SUA
}
\details{
This chunk compiles balanced supply utilization data bu connecting FAO datasets.
}
\author{
XZ 2023
}
